Set css_compressor in production to nil
This was still set on the purger from the tailwindcss-rails gem.
Since we now purge the tailwindcss directly we have to set this to nil.

The bundled sass parser is having issues with the tailwind-build.css output as
well.

Not sure yet if this is 100% the correct way to go. Time will tell :s)